refactor: restore global settings naming convention
