+++ frontend/src/components/settings/modal/sections/GlobalSettingsCodeSection.tsx
